pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pigeon-imessage"
version = "0.1.0"
description = "Turn iMessage into an AI assistant interface. Text your Mac, get LLM responses back."
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [{ name = "Greg Cosgrove" }]
keywords = ["imessage", "ai", "llm", "claude", "chatgpt", "ollama", "macos", "automation"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: MacOS X",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: MacOS",
    "Programming Language :: Python :: 3",
    "Topic :: Communications :: Chat",
]

dependencies = ["pyyaml>=6.0"]

[project.optional-dependencies]
anthropic = ["anthropic>=0.40"]
openai = ["openai>=1.50"]
postgres = ["psycopg2-binary>=2.9"]
all = ["anthropic>=0.40", "openai>=1.50", "psycopg2-binary>=2.9"]
dev = ["pytest>=8.0", "pytest-cov>=5.0", "ruff>=0.8"]

[project.scripts]
pigeon = "pigeon.cli:main"

[project.urls]
Homepage = "https://github.com/gbcosgrove/pigeon"
Issues = "https://github.com/gbcosgrove/pigeon/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/pigeon"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.pytest.ini_options]
testpaths = ["tests"]
